BackE/backend/requirements.txt
Django==3.1.2
django-allauth==0.42.0
django-cors-headers==3.5.0
django-rest-auth==0.9.5
requests==2.24.0
requests-oauthlib==1.3.0
asgiref==3.2.10
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
pytz==2020.1
sqlparse==0.4.1